+++ src/IRProtocol.h
@@ -71,36 +71,6 @@ typedef enum {
 
 } decode_type_t;
 
-const char string_Unknown[] PROGMEM = "UNKNOWN";
-const char string_PulseWidth[] PROGMEM = "PulseWidth";
-const char string_PulseDistance[] PROGMEM = "PulseDistance";
-const char string_Apple[] PROGMEM = "Apple";
-const char string_Denon[] PROGMEM = "Denon";
-const char string_JVC[] PROGMEM = "JVC";
-const char string_LG[] PROGMEM = "LG";
-const char string_LG2[] PROGMEM = "LG2";
-const char string_NEC[] PROGMEM = "NEC";
-const char string_NEC2[] PROGMEM = "NEC2";
-const char string_Onkyo[] PROGMEM = "Onkyo";
-const char string_Panasonic[] PROGMEM = "Panasonic";
-const char string_Kaseikyo[] PROGMEM = "Kaseikyo";
-const char string_Kaseikyo_Denon[] PROGMEM = "Kaseikyo_Denon";
-const char string_Kaseikyo_Sharp[] PROGMEM = "Kaseikyo_Sharp";
-const char string_Kaseikyo_JVC[] PROGMEM = "Kaseikyo_JVC";
-const char string_Kaseikyo_Mitsubishi[] PROGMEM = "Kaseikyo_Mitsubishi";
-const char string_RC5[] PROGMEM = "RC5";
-const char string_RC6[] PROGMEM = "RC6";
-const char string_Samsung[] PROGMEM = "Samsung";
-const char string_Samsung48[] PROGMEM = "Samsung48";
-const char string_SamsungLG[] PROGMEM = "SamsungLG";
-const char string_Sharp[] PROGMEM = "Sharp";
-const char string_Sony[] PROGMEM = "Sony";
-const char string_BangOlufsen[] PROGMEM = "Bang&Olufsen";
-const char string_BoseWave[] PROGMEM = "BoseWave";
-const char string_Lego[] PROGMEM = "Lego";
-const char string_MagiQuest[] PROGMEM = "MagiQuest";
-const char string_Whynter[] PROGMEM = "Whynter";
-
 struct PulseDistanceWidthProtocolConstants {
     decode_type_t ProtocolIndex;
     uint_fast8_t FrequencyKHz;