+++ changelog.md
@@ -32,7 +32,7 @@ See also the commit log at github: https://github.com/Arduino-IRremote/Arduino-I
 - Fixed SEND_PWM_BY_TIMER for ATtiny167 thanks to freskpe.
 - Improved SHARP repeat decoding.
 - Replaced macros TIMER_EN/DISABLE_RECEIVE_INTR and EN/DISABLE_SEND_PWM_BY_TIMER by functions.
-- Added SAMSUNG48 protocol and sendSamsung48().
+- Added SAMSUNG48 protocol and sendSamsung48() function.
 
 ## 3.9.0
 - Improved documentation with the help of [ElectronicsArchiver}(https://github.com/ElectronicsArchiver).